+++ internal/framework/services/s3/bucket_public_access_block_data_source.go
@@ -2,7 +2,6 @@ package s3
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
@@ -100,13 +99,10 @@ func (d *bucketPublicAccessBlockDataSource) Read(ctx context.Context, req dataso
 		return
 	}
 
-	response, err := GetBucketPublicAccessBlock(ctx, d.client, data.Bucket.ValueString())
+	bucket := data.Bucket.ValueString()
+	response, err := GetBucketPublicAccessBlock(ctx, d.client, bucket)
 	if err != nil {
-		if errors.Is(err, ErrBucketPublicAccessBlockNotFound) {
-			resp.State.RemoveResource(ctx)
-			return
-		}
-		resp.Diagnostics.AddError("Failed to retrieve bucket public access block", err.Error())
+		resp.Diagnostics.AddError(fmt.Sprintf("Failed to retrieve public access block for bucket: %s", bucket), err.Error())
 		return
 	}